backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pharos"
version = "0.1.0"
description = "A self-hosted, open-source Feedly alternative with LLM enrichment and deterministic story clustering."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Pharos contributors" }]
keywords = ["rss", "feed", "reader", "feedly", "llm", "threat-intelligence"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7.4",
  # passlib 1.7.4's bcrypt backend introspects bcrypt's private __about__
  # attribute and trips a 72-byte self-test on bcrypt>=4.x. Pin until
  # passlib cuts a release that supports the new bcrypt API.
  "bcrypt>=3.2,<4",
  "python-multipart>=0.0.9",
  "httpx>=0.27",
  "feedparser>=6.0.11",
  "trafilatura>=1.9",
  "apscheduler>=3.10",
  "openai>=1.30",
  "typer>=0.12",
  "rich>=13.7",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "respx>=0.21",
]

[project.scripts]
pharos = "pharos.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pharos*"]

[tool.setuptools.package-data]
"pharos.db"            = ["*.sql", "migrations/*.sql"]
"pharos.data"          = ["*.yaml", "*.yml"]
"pharos.lantern.data"  = ["*.json", "*.json.gz"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
